requirements.txt
Flask>=3.0
python-dateutil>=2.9
pytz>=2024.1
Jinja2>=3.1
click>=8.1
itsdangerous>=2.2
Werkzeug>=3.0
python-dotenv>=1.0
requests>=2.32
plaid-python>=16.0
gunicorn>=22.0
